requirements.txt
python-telegram-bot[job-queue]==20.7
python-dotenv==1.0.0
pyotp==2.9.0
pytz==2024.1
pytest==7.4.3
pytest-asyncio==0.23.2
